{{cookiecutter.repo_name}}/requirements/local.txt
# Includes base.txt requirements
-r base.txt

django-debug-toolbar==0.9.4